feat: move all conda types to seperate crate
