+++ README.md
@@ -5,4 +5,5 @@
 [build-badge]: https://img.shields.io/github/workflow/status/baszalmstra/rattler/Continuous%20integration?style=flat-square
 [build]: https://github.com/baszalmstra/rattler/actions
 
-Experiments to create Conda environments from Rust. Currently, not useful at all.
+Rattler is a suite of libraries that provides common Conda functionalities. 
+The goal of the libraries is to enable programs and other libraries to easily interact with the Conda ecosystem without being dependent on Python.
