+++ crates/rattler/src/repo_data/fetch/multi_request.rs
@@ -1,14 +1,12 @@
 //! Defines the [`MultiRequestRepoDataBuilder`] struct. This struct enables async fetching channel
 //! repodata from multiple source in parallel.
 
-use crate::utils::default_cache_dir;
-use crate::{
-    repo_data::fetch::request::{
-        RepoDataRequestState, RequestRepoDataBuilder, RequestRepoDataError, RequestRepoDataListener,
-    },
-    Channel, Platform, RepoData,
+use crate::repo_data::fetch::request::{
+    RepoDataRequestState, RequestRepoDataBuilder, RequestRepoDataError, RequestRepoDataListener,
 };
+use crate::utils::default_cache_dir;
 use futures::{stream::FuturesUnordered, StreamExt};
+use rattler_conda_types::{Channel, Platform, RepoData};
 use std::path::PathBuf;
 
 /// The `MultiRequestRepoDataBuilder` handles fetching data from multiple conda channels and
