+++ crates/rattler/src/repo_data/fetch/progress.rs
@@ -5,13 +5,14 @@ use std::{collections::HashMap, time::Duration};
 use indicatif::{MultiProgress, ProgressBar, ProgressDrawTarget, ProgressFinish, ProgressStyle};
 
 use super::{DoneState, DownloadingState, MultiRequestRepoDataListener, RepoDataRequestState};
-use crate::{Channel, Platform};
+use rattler_conda_types::{Channel, Platform};
 
 /// Returns a listener to use with the [`super::MultiRequestRepoDataBuilder`] that will show the
 /// progress as several progress bars.
 ///
 /// ```rust,no_run
 /// # use rattler::{Channel, repo_data::fetch::{ terminal_progress, MultiRequestRepoDataBuilder}, ChannelConfig};
+/// # use rattler_conda_types::{Channel, ChannelConfig};
 /// # tokio_test::block_on(async {
 /// let _ = MultiRequestRepoDataBuilder::default()
 ///     .add_channel(Channel::from_str("conda-forge", &ChannelConfig::default()).unwrap())
