+++ crates/rattler_conda_types/src/no_arch_type.rs
