+++ crates/rattler_conda_types/src/platform.rs
