+++ rust-toolchain
@@ -1,1 +1,1 @@
-1.63.0
+1.65.0
